fix(context-ledger): 压缩卡片布局并修正详情渲染
本次提交围绕 Context Ledger 面板做两类修复：
一是压缩最近轮次、比较摘要与来源卡片的视觉层级，减少无效文案与空白高度，提升主视图可读性与对齐稳定性；
二是修正来源详情的正文展示链路，避免致密单行 markdown、伪 section marker 与 feature 间耦合导致的内容错位、误切段和格式丢失。

主要改动：
- 重排 Context Ledger 面板结构，补充实时投影视图说明，压平卡片层级，收紧 comparison 摘要与 usage snapshot 的展示密度。
- 将 inspection title/content 扩展为 i18n key + params 模式，补齐中英文 copy，并把 recent turns 的详情改为更明确的实时用量说明。
- 新增 contextLedgerInspectionMarkdown 与 shared denseMarkdownOutput 归一化工具，恢复被压成单行的 markdown 结构，并限制 section parser 仅在行首或换行后识别标记，避免误拆普通正文。
- 让 session-activity 复用 shared markdown 归一化逻辑，消除 context-ledger 对 feature-local helper 的反向依赖。
- 增补回归测试，覆盖 labeled dense markdown 恢复、plain markdown 直通和 marker-like prose 不误切段。

验证：
- npm run lint
- npm run typecheck
- npm run check:large-files
- npx vitest run src/features/context-ledger/components/ContextLedgerPanel.test.tsx src/features/context-ledger/utils/contextLedgerProjection.test.ts src/features/context-ledger/utils/contextLedgerInspectionMarkdown.test.ts src/features/session-activity/utils/shellOutputHighlight.test.ts
- npm run check:heavy-test-noise
