+++ src/i18n/locales/en.part2.ts
@@ -1520,6 +1520,7 @@ const enPart2 = {
     contextLedgerSummaryBlocks_other: "{{count}} blocks",
     contextLedgerSummaryGroups: "{{count}} group",
     contextLedgerSummaryGroups_other: "{{count}} groups",
+    contextLedgerTruthNote: "Live projection · Built from the context snapshot used for this turn, not static help text.",
     contextLedgerGroupRecentTurns: "Recent turns",
     contextLedgerGroupCompaction: "Compaction",
     contextLedgerGroupManualMemory: "Manual memory",
@@ -1564,6 +1565,8 @@ const enPart2 = {
     contextLedgerTitleCompactionCompletedAuto: "Automatic compaction completed",
     contextLedgerDetailUsageWindow: "{{usedTokens}} / {{contextWindowTokens}} window · {{totalTokens}} total",
     contextLedgerDetailUsageTotal: "{{totalTokens}} total",
+    contextLedgerInspectionUsageWindow: "Current window {{usedTokens}} / {{contextWindowTokens}} · {{totalTokens}} historical tokens",
+    contextLedgerInspectionUsageTotal: "{{totalTokens}} historical tokens",
     contextLedgerDetailCompactionPendingSync: "Waiting for usage refresh",
     contextLedgerBackendSourceWorkspaceManaged: "Workspace managed",
     contextLedgerBackendSourceProjectClaude: "Project .claude",
@@ -1576,6 +1579,8 @@ const enPart2 = {
     contextLedgerBackendSourceGlobalAgents: "Global .agents",
     contextLedgerComparisonLastSend: "Compared with last send",
     contextLedgerComparisonPreCompaction: "Compared with pre-compaction snapshot",
+    contextLedgerComparisonLastSendHint: "Computed from ledger snapshots before and after the last send.",
+    contextLedgerComparisonPreCompactionHint: "Computed from ledger snapshots before and after compaction.",
     contextLedgerComparisonAdded: "Added {{count}}",
     contextLedgerComparisonRemoved: "Removed {{count}}",
     contextLedgerComparisonRetained: "Retained {{count}}",
