+++ src/i18n/locales/zh.part2.ts
@@ -1489,6 +1489,7 @@ const zhPart2 = {
     contextLedgerSummaryTokens: "{{tokens}} 标记",
     contextLedgerSummaryBlocks: "{{count}} 个条目",
     contextLedgerSummaryGroups: "{{count}} 个分组",
+    contextLedgerTruthNote: "实时投影 · 来自本轮将发送/刚使用的上下文快照，不是静态说明。",
     contextLedgerGroupRecentTurns: "最近轮次",
     contextLedgerGroupCompaction: "压缩状态",
     contextLedgerGroupManualMemory: "手动记忆",
@@ -1533,6 +1534,8 @@ const zhPart2 = {
     contextLedgerTitleCompactionCompletedAuto: "自动压缩已完成",
     contextLedgerDetailUsageWindow: "{{usedTokens}} / {{contextWindowTokens}} 窗口 · 总计 {{totalTokens}}",
     contextLedgerDetailUsageTotal: "总计 {{totalTokens}}",
+    contextLedgerInspectionUsageWindow: "当前窗口 {{usedTokens}} / {{contextWindowTokens}} · 历史总量 {{totalTokens}} 标记",
+    contextLedgerInspectionUsageTotal: "历史总量 {{totalTokens}} 标记",
     contextLedgerDetailCompactionPendingSync: "等待用量刷新",
     contextLedgerBackendSourceWorkspaceManaged: "工作区托管",
     contextLedgerBackendSourceProjectClaude: "项目 .claude",
@@ -1545,6 +1548,8 @@ const zhPart2 = {
     contextLedgerBackendSourceGlobalAgents: "全局 .agents",
     contextLedgerComparisonLastSend: "相比最近一次发送",
     contextLedgerComparisonPreCompaction: "相比压缩前快照",
+    contextLedgerComparisonLastSendHint: "基于上一次发送前后的账本快照计算。",
+    contextLedgerComparisonPreCompactionHint: "基于压缩前后的账本快照计算。",
     contextLedgerComparisonAdded: "新增 {{count}}",
     contextLedgerComparisonRemoved: "移除 {{count}}",
     contextLedgerComparisonRetained: "保留 {{count}}",
